Combined the search functions, pulled out utils
